Cell removal bug fix
The previous approach for fixing the bug, where i.js would not create
the new cells after evaluating the last one, relied on manipulating
nextCellId (see 4ba10de1db4f377e842d54d13b0c317e28a5768e). It did not
quite work in case if one removes the second to last cell and then the
last one. This fix is supposed to be more robust as it relies on the
analysis of DOM tree siblngs of the cell being evaluated. The previous
nextCellId decrement on the last cell removal was rolled back as
disfunctional.