Support multiline call chains
Before this fix an expression like:

var svg = chart.selectAll("div")
	.data(data)
	.enter()
	.append("div")
	.style("width", function(d) { return d * 10 + "px"; })
	.text(function(d) { return d; });

would not be properly interpreted: instead you would see an error
message “Invalid REPL keyword”. The reason for that is REPL is not
designed for dealing with multiline expressions and treats any line
starting with “.” character as a control command. This fix addresses it
by removing line feeds before chained method calls.